chore: quality pipeline finalization — GCC 15, ASan, memory leak fixes
Fix arena allocator alignment bug (align relative to actual address,
not block offset), add oas_doc_free() for yyjson_doc lifetime management,
store and free regex backend in compiled_doc. Fix GCC warnings:
-Wjump-misses-init, -Wformat-truncation, -Wconversion.

Verified: Clang debug 22/22, Clang ASan 21/21 (zero leaks),
GCC 15 debug 21/21, GCC 15 ASan 21/21. Quality pipeline 6/6 PASS.

+++ src/emitter/oas_emitter.c
@@ -63,9 +63,9 @@ static yyjson_mut_val *emit_type(yyjson_mut_doc *doc, uint8_t mask)
         return nullptr;
     }
 
-    for (uint8_t bit = 0x01; bit <= 0x40; bit <<= 1) {
+    for (unsigned bit = 0x01; bit <= 0x40; bit <<= 1) {
         if (mask & bit) {
-            const char *name = type_name_single(bit);
+            const char *name = type_name_single((uint8_t)bit);
             if (name) {
                 yyjson_mut_arr_append(arr, yyjson_mut_str(doc, name));
             }
